Game object mappings improvement. Game field table implement. Turned off emoji parser because of long init delay (3sec on emulator)

+++ app/src/main/java/com/fromfinalform/blocks/presentation/model/graphics/drawer/FlatShaderDrawer.kt
@@ -34,8 +34,8 @@ class FlatShaderDrawer() : IShaderDrawer {
         """
 
     override val FRAGMENT_SHADER = """
-        /*precision mediump float;*/
-        precision highp float;
+        precision mediump float;
+        /*precision highp float;*/
         varying vec2 vTextureCoord;
         uniform sampler2D sTexture;
         uniform float uAlpha;
